Add gnulib build script.
Also rework python tools directory.
