+++ batch/setup.bat
@@ -2,7 +2,8 @@
 
 set ROOT_DIR=%cd%\..
 set BATCH_TOOLS_DIR=%cd%\tools
-set PYTHON_TOOLS_DIR=%ROOT_DIR%\python\tools
+set PYTHON_TOOLS_DIR=%ROOT_DIR%\python\env\tools
+set PYTHON_LIB_DIR=%ROOT_DIR%\python\env\lib
 
 for /d  %%f in (%BATCH_TOOLS_DIR%\*) do (
     cd %%f
@@ -11,6 +12,12 @@ for /d  %%f in (%BATCH_TOOLS_DIR%\*) do (
 
 echo %PYTHON_TOOLS_DIR%
 for /d  %%f in (%PYTHON_TOOLS_DIR%\*) do (
+    cd %%f
+    python install.py
+)
+
+echo %PYTHON_LIB_DIR%
+for /d  %%f in (%PYTHON_LIB_DIR%\*) do (
     cd %%f
     python install.py
 )
