+++ python/vpy/git.py
@@ -2,7 +2,7 @@
 #
 # MIT License
 #
-# Copyright (c) 2021 Konychev Valera
+# Copyright (c) 2022 Konychev Valera
 #
 # Permission is hereby granted, free of charge, to any person obtaining a copy
 # of this software and associated documentation files (the "Software"), to deal
@@ -22,55 +22,11 @@
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 # SOFTWARE.
 
-import os
-import platform
+import subprocess
 
-OS = platform.system()
+GIT_BIN = 'git'
 
-def isLinux():
-    return OS == 'Linux'
-
-def isMacOsX():
-    return OS == 'Darwin'
-
-def isWin():
-    return OS == 'Windows'
-
-def execForOs(linux=None, mac=None, win=None):
-    if isLinux():
-        if linux is None:
-            raise 'Where is no Linux hadler'
-        linux()
-    elif isMacOsX():
-        if mac is None:
-            raise 'Where is no Mac OS X hadler'
-        mac()
-    elif isWin():
-        if win is None:
-            raise 'Where is no Win hadler'
-        win()
-    else:
-        raise 'Unknown OS'
-
-def getForOs(linux=None, mac=None, win=None):
-    if isLinux():
-        return linux
-    elif isMacOsX():
-        return mac
-    elif isWin():
-        return win
-    else:
-        raise 'Unknown OS'
-
-def appendPath(newPath):
-    def appendNixPath():
-        return
-    def appendWinPath():
-        os.system(f'setx path "%path%;{newPath}"')
-
-    execForOs(
-        linux = appendNixPath,
-        mac = appendNixPath,
-        win = appendWinPath
-    )
+def clone(args: [str]):
+    args = [GIT_BIN, 'clone'] + args
+    subprocess.run(args)
